+++ src/main.cpp
@@ -5,8 +5,10 @@
 #include "httplib.h"
 
 #include "task_manager.h"
+#include "save_manager.h"
 #include "request_manager.h"
 #include "file_utils.h"
+#include "settings.h"
 
 using std::cout, std::cerr, std::endl, std::string, std::shared_ptr, std::vector, std::atomic, std::map;
 using namespace httplib;
@@ -25,11 +27,7 @@ void signal_handler(int signum) {
 }
 
 int main(int argv, char* argc[] ) {
-    json config;
-    FileUtils::ParseConfigFile(DEFAULT_CONFIG_FILE, config);
-    string tasks_file = FileUtils::GetDefaultTasksFile(config);
-    string interface = FileUtils::GetDefaultInterface(config);
-    int port = FileUtils::GetDefaultPort(config);
+    Settings* settings = Settings::Instance();
 
     if (argv > 1) {
         if ((string)argc[1] == "-h") {
@@ -48,14 +46,14 @@ int main(int argv, char* argc[] ) {
                     cerr << argc[0] << " : Missing argument TASKS_FILE" << endl;
                     return 1;
                 }
-                tasks_file = argc[ii + 1];
+                settings->m_tasks_file = argc[ii + 1];
                 ii++;
             } else if ((string)argc[ii] == "-i") {
                 if (ii == argv - 1) {
                     cerr << argc[0] << " : Missing argument INTERFACE_ADDR" << endl;
                     return 1;
                 }
-                interface = argc[ii + 1];
+                settings->m_interface = argc[ii + 1];
                 ii++;
             } else if ((string)argc[ii] == "-p") {
                 if (ii == argv - 1) {
@@ -63,7 +61,7 @@ int main(int argv, char* argc[] ) {
                     return 1;
                 }
                 try {
-                    port = std::stoi(argc[ii + 1]);
+                    settings->m_port = std::stoi(argc[ii + 1]);
                 } catch (...) {
                     cerr << argc[0] << " : Invalid argument PORT" << argc[ii + 1] << endl;
                     return 1;
@@ -80,11 +78,17 @@ int main(int argv, char* argc[] ) {
     // Setup TaskManager
     shared_ptr<TaskManager> tm;
     try {
-        tm = std::make_shared<TaskManager>(tasks_file);
+        tm = std::make_shared<TaskManager>(settings->m_tasks_file);
     } catch (const std::exception& e) {
         // failed to initialize
         cerr << e.what();
         return 1;
+    }
+
+    // only setup periodic saving if save period is greater than 0 seconds (float)
+    if (settings->m_save_period >= 1e-9) {
+        // Setup SaveManager
+        shared_ptr<SaveManager> sm = std::make_shared<SaveManager>(tm);
     }
 
     // Setup RequestManager
@@ -97,16 +101,17 @@ int main(int argv, char* argc[] ) {
     // Signal catching
     std::signal(SIGINT, signal_handler);
     std::signal(SIGTERM, signal_handler);
+    std::signal(SIGHUP, signal_handler);
 
     std::thread server_thread;
     try {
         // start http server on seperate thread
         server_thread = std::thread([&]() {
-            cout << "Server listening on " << interface << ":" << port << "..." << endl;
-            svr.listen(interface, port);
+            cout << "Server listening on " << settings->m_interface << ":" << settings->m_port << "..." << endl;
+            svr.listen(settings->m_interface, settings->m_port);
         });
     } catch (...) {
-        cerr << argc[0] << " : Failed to open server on " << interface << ":" << port << endl;
+        cerr << argc[0] << " : Failed to open server on " << settings->m_interface << ":" << settings->m_port << endl;
         return 1;
     }
 
@@ -118,6 +123,8 @@ int main(int argv, char* argc[] ) {
     // stop server and join its thread
     svr.stop();
     server_thread.join();
+    // save tasks to disk
+    tm->Save();
     cout << "Server stopped." << endl;
     
     return 0;